Moved config to src, comp errors.
