+++ src/config/calendar.ts
@@ -1,3 +1,4 @@
+
 export class BossyCalendarConfig {
     defaultDate: Date;
